Introduce method to compare only organisation part of Organisasjonsnummer
The changes introduced in the two previous commits change the behaviour
of the Organisasjonsnummer equals method. Previously, this method could
be used to compare only the organisation number string itself between two
instances of Organisasjonsnummer. The method introduced in this commit,
erSammeOrganisasjonsnummerUavhengigAvLandkode, is intended to replace
the described use of the equals method for uses where ignoring country
code is desirable.